[DDD light] Checklists & checklist items now go hand in hand with the new architecture

+++ backend/ArhiTodo.Application/Services/Implementations/Kanban/ChecklistService.cs
@@ -4,16 +4,22 @@ using ArhiTodo.Application.Mappers;
 using ArhiTodo.Application.Services.Interfaces.Kanban;
 using ArhiTodo.Application.Services.Interfaces.Realtime;
 using ArhiTodo.Domain.Entities.Kanban;
+using ArhiTodo.Domain.Repositories.Common;
 using ArhiTodo.Domain.Repositories.Kanban;
 
 namespace ArhiTodo.Application.Services.Implementations.Kanban;
 
-public class ChecklistService(IChecklistNotificationService checklistNotificationService, IChecklistRepository checklistRepository) : IChecklistService
+public class ChecklistService(ICardRepository cardRepository, IChecklistNotificationService checklistNotificationService,
+    IUnitOfWork unitOfWork) : IChecklistService
 {
     public async Task<ChecklistGetDto?> CreateChecklist(int boardId, int cardId, ChecklistCreateDto checklistCreateDto)
     {
-        Checklist? checklist = await checklistRepository.CreateChecklistOnCard(checklistCreateDto.FromCreateDto(cardId));
-        if (checklist == null) return null;
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return null;
+
+        Checklist checklist = new(cardId, checklistCreateDto.ChecklistName);
+        card.AddChecklist(checklist);
+        await unitOfWork.SaveChangesAsync();
 
         ChecklistGetDto checklistGetDto = checklist.ToGetDto();
         checklistNotificationService.CreateChecklist(boardId, cardId, checklistGetDto);
@@ -22,61 +28,98 @@ public class ChecklistService(IChecklistNotificationService checklistNotificatio
 
     public async Task<ChecklistGetDto?> UpdateChecklist(int boardId, int cardId, ChecklistUpdateDto checklistUpdateDto)
     {
-        Checklist? checklist = await checklistRepository.UpdateChecklist(checklistUpdateDto.FromUpdateDto(cardId));
-        if (checklist == null) return null;
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return null;
 
+        Checklist? checklist = card.Checklists.FirstOrDefault(cl => cl.ChecklistId == checklistUpdateDto.ChecklistId);
+        if (checklist == null) return null;
+        
+        checklist.RenameChecklist(checklistUpdateDto.ChecklistName);
+        await unitOfWork.SaveChangesAsync();
+        
         ChecklistGetDto checklistGetDto = checklist.ToGetDto();
         checklistNotificationService.UpdateChecklist(boardId, cardId, checklistGetDto);
         return checklistGetDto;
     }
 
-    public async Task<bool> DeleteChecklist(int boardId, int checklistId)
+    public async Task<bool> DeleteChecklist(int boardId, int cardId, int checklistId)
     {
-        bool succeeded = await checklistRepository.DeleteChecklistFromCard(checklistId);
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return false;
+
+        bool succeeded = card.RemoveChecklist(checklistId);
         if (succeeded)
         {
             checklistNotificationService.DeleteChecklist(boardId, checklistId);
+            await unitOfWork.SaveChangesAsync();
         }
         return succeeded;
     }
 
-    public async Task<ChecklistItemGetDto?> CreateChecklistItem(int boardId, int checklistId, ChecklistItemCreateDto checklistItemCreateDto)
+    public async Task<ChecklistItemGetDto?> CreateChecklistItem(int boardId, int cardId, int checklistId, 
+        ChecklistItemCreateDto checklistItemCreateDto)
     {
-        ChecklistItem? createdChecklistItem = await checklistRepository.AddChecklistItemToChecklist(checklistItemCreateDto.FromCreateDto(checklistId));
-        if (createdChecklistItem == null) return null;
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return null;
+
+        Checklist? checklist = card.Checklists.FirstOrDefault(c => c.ChecklistId == checklistId);
+        if (checklist == null) return null;
+        
+        ChecklistItem createdChecklistItem = checklist.AddChecklistItem(checklistItemCreateDto.ChecklistItemName);
+        await unitOfWork.SaveChangesAsync();
 
         ChecklistItemGetDto checklistItemGetDto = createdChecklistItem.ToGetDto();
         checklistNotificationService.CreateChecklistItemOnChecklist(boardId, checklistId, checklistItemGetDto);
         return checklistItemGetDto;
     }
-
-    public async Task<ChecklistItemGetDto?> UpdateChecklistItem(int boardId, int checklistId, ChecklistItemUpdateDto checklistItemUpdateDto)
+    
+    public async Task<bool> DeleteChecklistItem(int boardId, int cardId, int checklistId, int checklistItemId)
     {
-        ChecklistItem? updatedChecklistItem =
-            await checklistRepository.UpdateChecklistItem(checklistItemUpdateDto.FromUpdateDto(checklistId));
-        if (updatedChecklistItem == null) return null;
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return false;
+        
+        Checklist? checklist = card.Checklists.FirstOrDefault(c => c.ChecklistId == checklistId);
+        if (checklist == null) return false;
 
-        ChecklistItemGetDto checklistItemGetDto = updatedChecklistItem.ToGetDto();
-        checklistNotificationService.UpdateChecklistItem(boardId, checklistId, checklistItemGetDto);
-        return checklistItemGetDto;
-    }
-
-    public async Task<bool> DeleteChecklistItem(int boardId, int checklistId, int checklistItemId)
-    {
-        bool succeeded = await checklistRepository.RemoveChecklistItemFromChecklist(checklistItemId);
+        bool succeeded = checklist.RemoveChecklistItem(checklistItemId);
         if (succeeded)
         {
+            await unitOfWork.SaveChangesAsync();
             checklistNotificationService.DeleteChecklistItemFromChecklist(boardId, checklistId, checklistItemId);
         }
         return succeeded;
     }
 
-    public async Task<ChecklistItemGetDto?> PatchChecklistItemState(int boardId, int checklistItemId, bool newState)
+    public async Task<ChecklistItemGetDto?> UpdateChecklistItem(int boardId, int cardId, int checklistId, 
+        ChecklistItemUpdateDto checklistItemUpdateDto)
     {
-        ChecklistItem? checklistItem = await checklistRepository.PatchChecklistItemDoneState(checklistItemId, newState);
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return null;
+        
+        Checklist? checklist = card.Checklists.FirstOrDefault(c => c.ChecklistId == checklistId);
+        if (checklist == null) return null;
+        ChecklistItem? checklistItem = checklist.ChecklistItems.FirstOrDefault(ci => ci.ChecklistItemId == checklistItemUpdateDto.ChecklistItemId);
         if (checklistItem == null) return null;
+        
+        checklistItem.RenameChecklistItem(checklistItemUpdateDto.ChecklistItemName);
+        await unitOfWork.SaveChangesAsync();
 
         ChecklistItemGetDto checklistItemGetDto = checklistItem.ToGetDto();
+        checklistNotificationService.UpdateChecklistItem(boardId, checklistId, checklistItemGetDto);
+        return checklistItemGetDto;
+    }
+
+    public async Task<ChecklistItemGetDto?> PatchChecklistItemState(int boardId, int cardId, int checklistItemId, bool newState)
+    {
+        Card? card = await cardRepository.GetDetailedCard(cardId);
+        if (card == null) return null;
+
+        ChecklistItem? updateChecklistItem = card.UpdateChecklistItemState(checklistItemId, newState);
+        if (updateChecklistItem == null) return null;
+
+        await unitOfWork.SaveChangesAsync();
+        
+        ChecklistItemGetDto checklistItemGetDto = updateChecklistItem.ToGetDto();
         checklistNotificationService.PatchChecklistItemDoneState(boardId, checklistItemId, newState);
         return checklistItemGetDto;
     }
